requirements.txt
fastapi
SQLAlchemy
python-dotenv
python-jose
passlib
uvicorn
psycopg2
livekit-api